template: remove sandboxing on windows
