+++ client/allocrunner/taskrunner/template/template.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"log"
 	"math/rand"
 	"os"
 	"path/filepath"
@@ -19,17 +18,13 @@ import (
 
 	ctconf "github.com/hashicorp/consul-template/config"
 	"github.com/hashicorp/consul-template/manager"
-	"github.com/hashicorp/consul-template/renderer"
-	"github.com/hashicorp/consul-template/signals"
 	envparse "github.com/hashicorp/go-envparse"
 	"github.com/hashicorp/go-hclog"
 	multierror "github.com/hashicorp/go-multierror"
 	"github.com/hashicorp/nomad/client/allocrunner/taskrunner/interfaces"
-	trenderer "github.com/hashicorp/nomad/client/allocrunner/taskrunner/template/renderer"
 	"github.com/hashicorp/nomad/client/config"
 	"github.com/hashicorp/nomad/client/taskenv"
 	"github.com/hashicorp/nomad/helper/pointer"
-	"github.com/hashicorp/nomad/helper/subproc"
 	"github.com/hashicorp/nomad/nomad/structs"
 	structsc "github.com/hashicorp/nomad/nomad/structs/config"
 )
@@ -163,74 +158,6 @@ func (c *TaskTemplateManagerConfig) Validate() error {
 	}
 
 	return nil
-}
-
-func NewTaskTemplateManager(config *TaskTemplateManagerConfig) (*TaskTemplateManager, error) {
-	// Check pre-conditions
-	if err := config.Validate(); err != nil {
-		return nil, err
-	}
-
-	tm := &TaskTemplateManager{
-		config:     config,
-		shutdownCh: make(chan struct{}),
-	}
-
-	// Parse the signals that we need
-	for _, tmpl := range config.Templates {
-		if tmpl.ChangeSignal == "" {
-			continue
-		}
-
-		sig, err := signals.Parse(tmpl.ChangeSignal)
-		if err != nil {
-			return nil, fmt.Errorf("Failed to parse signal %q", tmpl.ChangeSignal)
-		}
-
-		if tm.signals == nil {
-			tm.signals = make(map[string]os.Signal)
-		}
-
-		tm.signals[tmpl.ChangeSignal] = sig
-	}
-
-	// the platform sandbox needs to be created before we construct the runner
-	// so that reading the template is sandboxed
-	err := createPlatformSandbox(config)
-	if err != nil {
-		return nil, err
-	}
-
-	// Build the consul-template runner
-	runner, lookup, err := templateRunner(config)
-	if err != nil {
-		return nil, err
-	}
-	tm.runner = runner
-	tm.lookup = lookup
-
-	go tm.run()
-	return tm, nil
-}
-
-// Stop is used to stop the consul-template runner
-func (tm *TaskTemplateManager) Stop() {
-	tm.shutdownLock.Lock()
-	defer tm.shutdownLock.Unlock()
-
-	if tm.shutdown {
-		return
-	}
-
-	close(tm.shutdownCh)
-	tm.shutdown = true
-
-	// Stop the consul-template runner
-	if tm.runner != nil {
-		tm.runner.Stop()
-	}
-
-	destroyPlatformSandbox(tm.config)
 }
 
 // SetDriverHandle sets the executor
@@ -993,109 +920,6 @@ func newRunnerConfig(config *TaskTemplateManagerConfig,
 	conf.RendererFunc = RenderFn(config.TaskID, sandboxDir, sandboxEnabled)
 	conf.Finalize()
 	return conf, nil
-}
-
-func isSandboxEnabled(cfg *TaskTemplateManagerConfig) bool {
-	if cfg.ClientConfig != nil && cfg.ClientConfig.TemplateConfig != nil && cfg.ClientConfig.TemplateConfig.DisableSandbox {
-		return false
-	}
-	return true
-}
-
-type sandboxConfig struct {
-	thisBin     string
-	sandboxPath string
-	destPath    string
-	sourcePath  string
-	perms       string
-	user        string
-	group       string
-	taskID      string
-	contents    []byte
-}
-
-func ReaderFn(taskID, taskDir string, sandboxEnabled bool) func(string) ([]byte, error) {
-	if !sandboxEnabled {
-		return nil
-	}
-	thisBin := subproc.Self()
-
-	return func(src string) ([]byte, error) {
-
-		sandboxCfg := &sandboxConfig{
-			thisBin:     thisBin,
-			sandboxPath: taskDir,
-			sourcePath:  src,
-			taskID:      taskID,
-		}
-
-		stdout, stderr, code, err := readTemplateFromSandbox(sandboxCfg)
-		if err != nil && code != 0 {
-			return nil, fmt.Errorf("%v: %s", err, string(stderr))
-		}
-
-		// this will get wrapped in CT log formatter
-		fmt.Fprintf(os.Stderr, "[DEBUG] %s", string(stderr))
-		return stdout, nil
-	}
-}
-
-func RenderFn(taskID, taskDir string, sandboxEnabled bool) func(*renderer.RenderInput) (*renderer.RenderResult, error) {
-	if !sandboxEnabled {
-		return nil
-	}
-	thisBin := subproc.Self()
-
-	return func(i *renderer.RenderInput) (*renderer.RenderResult, error) {
-		wouldRender := false
-		didRender := false
-
-		sandboxCfg := &sandboxConfig{
-			thisBin:     thisBin,
-			sandboxPath: taskDir,
-			destPath:    i.Path,
-			perms:       strconv.FormatUint(uint64(i.Perms), 8),
-			user:        i.User,
-			group:       i.Group,
-			taskID:      taskID,
-			contents:    i.Contents,
-		}
-
-		logs, code, err := renderTemplateInSandbox(sandboxCfg)
-		if err != nil {
-			if len(logs) > 0 {
-				log.Printf("[ERROR] %v: %s", err, logs)
-			} else {
-				log.Printf("[ERROR] %v", err)
-			}
-			return &renderer.RenderResult{
-				DidRender:   false,
-				WouldRender: false,
-				Contents:    []byte{},
-			}, fmt.Errorf("template render subprocess failed: %w", err)
-		}
-		if code == trenderer.ExitWouldRenderButDidnt {
-			didRender = false
-			wouldRender = true
-		} else {
-			didRender = true
-			wouldRender = true
-		}
-
-		// the subprocess emits logs matching the consul-template runner, but we
-		// CT doesn't support hclog, so we just print the whole output here to
-		// stderr the same way CT does so the results look seamless
-		if len(logs) > 0 {
-			log.Printf("[DEBUG] %s", logs)
-		}
-
-		result := &renderer.RenderResult{
-			DidRender:   didRender,
-			WouldRender: wouldRender,
-			Contents:    i.Contents,
-		}
-		return result, nil
-	}
 }
 
 // loadTemplateEnv loads task environment variables from all templates.
